refactor: extract tools.js from main.js (BAT-204)
Largest extraction in the modular refactor — moves ~3,500 lines from
main.js into tools.js: TOOLS array, executeTool() dispatch switch,
confirmation flow, and helper functions.

- tools.js (3,523 lines): TOOLS[], executeTool(), pendingConfirmations,
  requestConfirmation, formatConfirmationMessage, listFilesRecursive,
  formatBytes, setMcpExecuteTool() injection
- main.js: 4,238 → 767 lines (removed 3,471 lines)
- Cleaned up 20+ unused imports across all module import blocks
- Fixed pre-existing bug: WORKSPACE_DIR → MEMORY_DIR in /status command
